Update react example to newest versions
